pyproject.toml
[tool.poetry]
name = "cognition-engine"
version = "0.1.0.dev15"
description = "IOC CFN Cognition Engine – unified knowledge extraction, evidence gathering and semantic negotiation."
authors = ["Your Team <team@example.com>"]
readme = "README.md"
license = "Apache-2.0"
homepage = "https://github.com/cisco-eti/ioc-cfn-cognitive-agents"
repository = "https://github.com/cisco-eti/ioc-cfn-cognitive-agents"
keywords = ["knowledge-graph", "llm", "agents", "opentelemetry"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]
package-mode = true
packages = [
    { include = "ingestion" },
    { include = "evidence" },
    { include = "caching" },
    { include = "gateway" },
    { include = "protocol" },
    { include = "semantic_negotiation" },
]
exclude = [
    "test_callback_agents_go",
    "test_callback_agents_go/**",
    "sample_trace",
    "sample_trace/**",
    "neg_trace",
    "neg_trace/**",
    "semantic_negotiation/neg_trace",
    "semantic_negotiation/neg_trace/**",
    "semantic_negotiation/tests",
    "semantic_negotiation/tests/**",
    "semantic_negotiation/test_*.py",
    "semantic_negotiation/*.sh",
    "semantic_negotiation/.env",
    "semantic_negotiation/evaluation",
    "semantic_negotiation/evaluation/**",
]

# ── Shared runtime dependencies (union of all four services) ─────────────────
[tool.poetry.dependencies]
python = "^3.11"

# Web framework & ASGI server
fastapi = ">=0.110"
uvicorn = {version = ">=0.27", extras = ["standard"]}
httpx = ">=0.27"

# Data validation / settings
pydantic = ">=2.12.5"
pydantic-settings = ">=2.0.0"
python-dotenv = ">=1.0.1"

# AI / ML – fastembed is a lightweight ONNX-based embedding library;
# no PyTorch / CUDA dependency, image stays small.
fastembed = ">=0.4.0"
typer = ">=0.9.0"
openai = ">=2.17.0"
boto3 = ">=1.28.0"
numpy = ">=1.26,<3"

# Vector store
faiss-cpu = ">=1.7.4"

# Negotiation framework (semantic_negotiation)
negmas = ">=0.10"

# ── Security pins (CVEs) ──────────────────────────────────────────────────────
urllib3 = ">=2.6.3"   # CVE-2026-21441: decompression-bomb bypass on redirects
# Pillow: cannot pin >=12.1.1 here – fastembed requires pillow<12.0. CVE-2026-25990
# (PSD out-of-bounds write) remains until fastembed supports Pillow 12+. Mitigate by
# not opening untrusted PSDs or use Image.open(..., formats=...) to exclude PSD.

# ── Service-specific extras kept at top level ─────────────────────────────────
# (langchain-openai, pandas, requests, werkzeug removed – not imported in any
#  app/*.py file; were inherited from an older pyproject.toml template)

# ── Dev / test dependencies ───────────────────────────────────────────────────
[tool.poetry.group.dev.dependencies]
pytest = ">=7.0.0"
pytest-cov = ">=4.1.0"
pytest-asyncio = ">=0.23.0"
httpx = ">=0.27"          # also needed at runtime but explicit here for tests
ruff = ">=0.1.0"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
testpaths = [
    "gateway/tests",
    "ingestion/tests",
    "evidence/tests",
    "caching/tests",
    "semantic_negotiation/tests",
    "tests",
]
python_files = "test_*.py"
python_classes = "Test*"
python_functions = "test_*"
addopts = "-v -s --tb=short -m 'not requires_credentials'"
markers = [
    "requires_credentials: marks tests that require Azure OpenAI credentials (skipped by default in CI)",
]

[tool.ruff]
line-length = 120
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "W"]
ignore = ["E501"]
